+++ volto.config.js
@@ -1,4 +1,4 @@
-const addons = ['@redturtle/volto-blocks'];
+const addons = ['@redturtle/volto-slate-extras', '@redturtle/volto-blocks'];
 const theme = '';
 
 module.exports = {